% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anae}
\alias{anae}
\title{Dialect regions: ANAE}
\format{
An object of class \code{"sf", "data.frame"}.
}
\source{
YGDP GIS work
}
\usage{
data(anae)
}
\description{
Dialect region polygons for regions as defined in the Atlas of North American English.
}
\examples{
data(anae)

}
\keyword{datasets}
