% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carver}
\alias{carver}
\title{Dialect regions: Carver}
\format{
An object of class \code{"sf", "data.frame"}.
}
\source{
YGDP GIS work
}
\usage{
data(carver)
}
\description{
Dialect region polygons following Carver (1987), American Regional Dialects.
}
\examples{
data(carver)

}
\keyword{datasets}
