% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCompletion.R
\name{checkCompletion}
\alias{checkCompletion}
\title{Check survey completion and annotate the AMT table}
\usage{
checkCompletion(amt, qt, n = 2)
}
\arguments{
\item{amt}{The raw AMT output data frame.}

\item{qt}{The (somewhat cleaned) Qualtrics output, after having passed through `preprocessData()`.}

\item{n}{The number of un-rated sentences acceptable in a "complete" survey. Default is 2, although settings on newer surveys to force answers to questions, this may actually be too liberal.}
}
\description{
This function takes three inputs: the amt output table, the (cleaned) raw survey data table, and a parameter to define how many un-rated sentences are acceptable (defaults to 2). For each response, the function computes how many unanswered questions that person had and marks the survey as complete or not based on that and the parameter. Then it joins that result, by completion code, to the AMT table.
}
