% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFunctions.R
\name{checkMasterListColumns}
\alias{checkMasterListColumns}
\title{Check whether the master list contains the required columns}
\usage{
checkMasterListColumns(masterList)
}
\arguments{
\item{masterList}{The master list of YGDP sentences.}
}
\description{
In order for the survey processing functions to work, the master list must contain columns "sentenceID", "sentenceText", and "constructionID". This function checks that all columns are present and throws an informative error if not, telling you which ones are missing. If all columns are present, the function gives you a SUCCESS message.
}
