% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFunctions.R
\name{checkMasterListSentenceIDs}
\alias{checkMasterListSentenceIDs}
\title{Check whether the master list has any duplicate sentenceID's in the 'sentenceID' column.}
\usage{
checkMasterListSentenceIDs(masterList)
}
\arguments{
\item{masterList}{The master list of YGDP sentences.}
}
\description{
In order for joins to work properly in the survey processing script, the master list must not have any duplicate sentenceID's. This function throws informative errors if the "sentenceID" column is missing or if there are duplicate sentenceID's, and prints a success message if no duplicates are found.
}
