% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFunctions.R
\name{checkRawDupNames}
\alias{checkRawDupNames}
\title{Check for duplicate column-names-to-be in the raw data}
\usage{
checkRawDupNames(rawData)
}
\arguments{
\item{rawData}{The raw Qualtrics output.}
}
\description{
This function checks the first row of the raw data (which will become the column names after cleanup) to see if there are any duplicate column-names-to-be. If there are, it throws an error, and you'll need to go and fix those duplicates.
}
