% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFunctions.R
\name{checkRawSentenceNames}
\alias{checkRawSentenceNames}
\title{Check how many TS, CU, CG, and comments columns the data contains}
\usage{
checkRawSentenceNames(rawData)
}
\arguments{
\item{rawData}{The raw Qualtrics output.}
}
\description{
This function checks the first row of the raw data, which will become the column names after cleaning. It reports how many sentence columns (marked with "TS_", "CU_", and "CG_") it finds, as well as how many comments columns (marked with "comments"). If the number of sentence columns doesn't match the number of comments columns, the function throws a warning.
}
