% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFunctions.R
\name{checkSentenceCommentNumbers}
\alias{checkSentenceCommentNumbers}
\title{Check that the number of comments columns matches the number of sentence columns}
\usage{
checkSentenceCommentNumbers(rawData)
}
\arguments{
\item{rawData}{The raw Qualtrics output.}
}
\description{
This function checks that the number of columns whose name-to-be (i.e. the first row of the raw data) starts with "comments_" matches the number that start with "TS_", "CU_", or "CG_", as per the instructions in the survey format guidelines document. If there's a mismatch, the function throws a warning that you may end up with some missing data.
}
