% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixNames.R
\name{fixNames}
\alias{fixNames}
\title{Fix the names from the qualtrics table}
\usage{
fixNames(df)
}
\arguments{
\item{df}{Raw survey data as a data frame. Must have column 'responseID'.}

\item{updateID}{Character string that will become the `updateID` column for this table. For example, "survey11Add" for Survey 11.}
}
\description{
Fix the names from the qualtrics table
}
