% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkFunctions.R
\name{howManySentences}
\alias{howManySentences}
\title{Check how many TS, CU, CG, and comments columns the data contains}
\usage{
howManySentences(rawData)
}
\arguments{
\item{rawData}{The raw Qualtrics output.}
}
\description{
This function checks the first row of the raw data, which will become the column names after cleaning. It reports how many sentence columns (marked with "TS_", "CU_", and "CG_") it finds, as well as how many comments columns (marked with "comments"). You need to manually look at these numbers to make sure they make sense.
}
