% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTables.R
\name{makeCensusUrbanAreas}
\alias{makeCensusUrbanAreas}
\title{Make the CENSUS_URBAN_AREAS table}
\usage{
makeCensusUrbanAreas(cities, con, updateID)
}
\arguments{
\item{cities}{The CITIES element of the list returned by makeCensusCountyDemo.R (i.e. its final form)}

\item{con}{A connection to the existing database, created with RSQLite.}

\item{updateID}{Character string that will become the `updateID` column for this table. For example, "survey11Add" for Survey 11.}
}
\description{
This function returns a data frame, the CENSUS_URBAN_AREAS table of the database.
}
