% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTables.R
\name{makeComments}
\alias{makeComments}
\title{Make the COMMENTS table}
\usage{
makeComments(df, questions, updateID)
}
\arguments{
\item{df}{Raw survey data as a data frame.}

\item{questions}{The QUESTIONS portion of the database tables, created by makeQuestions.R}

\item{updateID}{Character string that will become the `updateID` column for this table. For example, "survey11Add" for Survey 11.}
}
\description{
This function creates the COMMENTS database table from raw survey data.
}
