% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeConstructions.R
\name{makeConstructions}
\alias{makeConstructions}
\title{Make the CONSTRUCTIONS table}
\usage{
makeConstructions(constructions, con, updateID)
}
\arguments{
\item{constructions}{A csv file, which you need to create separately before running this script. The table should have three columns: constructionID (existing or new ID's for all of the constructions contained in this survey), constructionName (the name of each construction, corresponding to the ID's), and pageYGDP (the URL of the YGDP page corresponding to this construction, if applicable.)}

\item{con}{A connection to the existing database, created with RSQLite.}

\item{updateID}{Character string that will become the `updateID` column for this table. For example, "survey11Add" for Survey 11.}
}
\description{
Make the CONSTRUCTIONS table
}
