% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTables.R
\name{makeDemoGeo}
\alias{makeDemoGeo}
\title{Make the DEMO_GEO table}
\usage{
makeDemoGeo(df, updateID, key, con, overwrite = T)
}
\arguments{
\item{df}{Raw survey data as a data frame.}

\item{updateID}{Character string that will become the `updateID` column for this table. For example, "survey11Add" for Survey 11.}

\item{key}{The API key for the hereR API}

\item{con}{A connection to the existing database, created with RSQLite.}

\item{overwrite}{Whether or not to overwrite existing geocoded localities, if `geocoded` already exists in the environment. This parameter was mainly included for function testing and development--it pretty much never makes sense to set it to F when using the function.}
}
\description{
This function returns a list, not a data frame. The list contains the DEMO_GEO, CITIES, and CITIES_REF portions of the database. Note that for CITIES to be complete, it has to be further passed through the makeCensusCountyDemo.R function, which also outputs a list.
}
