% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTables.R
\name{makeQuestions}
\alias{makeQuestions}
\title{Make the QUESTIONS table}
\usage{
makeQuestions(df, qids, surveyID, updateID)
}
\arguments{
\item{df}{Raw survey data as a data frame.}

\item{qids}{Data frame containing questionID/column name mappings for this survey. This is created using the `saveQuestionIDs()` function. Must have columns 'colName' and 'questionIDRaw'.}

\item{surveyID}{Character string that will become the `surveyID` column for this table. For example, "S11" for Survey 11.}

\item{updateID}{Character string that will become the `updateID` column for this table. For example, "survey11Add" for Survey 11.}
}
\description{
This function creates the QUESTIONS database table from raw survey data.
}
