% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTables.R
\name{makeRatings}
\alias{makeRatings}
\title{Make the RATINGS table}
\usage{
makeRatings(df, questions, surveyID, updateID)
}
\arguments{
\item{df}{Raw survey data as a data frame.}

\item{questions}{The QUESTIONS portion of the database tables, created by makeQuestions.R}

\item{surveyID}{Character string that will become the `surveyID` column for this table. For example, "S11" for Survey 11.}

\item{updateID}{Character string that will become the `updateID` column for this table. For example, "survey11Add" for Survey 11.}
}
\description{
This function creates the RATINGS database table from raw survey data.
}
