% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTables.R
\name{makeResponses}
\alias{makeResponses}
\title{Make the RESPONSES table}
\usage{
makeResponses(df, updateID, surveyID)
}
\arguments{
\item{df}{Raw survey data as a data frame. Must have column 'responseID'.}

\item{updateID}{Character string that will become the `updateID` column for this table. For example, "survey11Add" for Survey 11.}

\item{surveyID}{Character string that will become the `surveyID` column for this table. For example, "S11" for Survey 11.}
}
\description{
This function creates the RESPONSES database table from raw survey data.
}
