% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSentences.R
\name{makeSentences}
\alias{makeSentences}
\title{Make the SENTENCES table}
\usage{
makeSentences(df, masterList, con, updateID)
}
\arguments{
\item{df}{Raw survey data as a data frame.}

\item{masterList}{Master list of all sentences. ### XXX COME BACK TO THIS: HOW BEST TO LOAD AND UPDATE?}

\item{con}{Connection to the current version of the database. Create the connection using RSQLite. For example: con <- dbConnect(RSQLite::SQLite(), here("database", "currentDB", "ygdpDB.db")) (for the database file ygdpDB.db, stored in "database/currentDB".)}

\item{updateID}{Character string that will become the `updateID` column for this table. For example, "survey11Add" for Survey 11.}
}
\description{
This function creates the SENTENCES database table from raw survey data.
}
