% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSpokenLangs.R
\name{makeSpokenLangs}
\alias{makeSpokenLangs}
\title{Make the SPOKEN_LANGS table}
\usage{
makeSpokenLangs(df, updateID)
}
\arguments{
\item{df}{Raw survey data as a data frame. Must have columns 'responseID', 'consent', and 'completionCode'.}

\item{updateID}{Character string that will become the `updateID` column for this table. For example, "survey11Add" for Survey 11.}
}
\description{
Make the SPOKEN_LANGS table
}
