% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSurveyComments.R
\name{makeSurveyComments}
\alias{makeSurveyComments}
\title{Make the SURVEY_COMMENTS table}
\usage{
makeSurveyComments(df, surveyID, qids, updateID)
}
\arguments{
\item{df}{Raw survey data as a data frame. Must have columns 'responseID' and 'generalComments'.}

\item{surveyID}{Character string that will become the `surveyID` column for this table. For example, "S11" for Survey 11.}

\item{qids}{Data frame containing questionID/column name mappings for this survey. Create this using the `saveQuestionIDs()` function.}

\item{updateID}{Character string that will become the `updateID` column for this table. For example, "survey11Add" for Survey 11.}
}
\description{
Make the SURVEY_COMMENTS table
}
