% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTables.R
\name{makeTech}
\alias{makeTech}
\title{This function creates the TECH database table from raw survey data.}
\usage{
makeTech(df, updateID)
}
\arguments{
\item{df}{Raw survey data as a data frame. Must have columns 'responseID', 'consent', and 'completionCode'.}

\item{updateID}{Character string that will become the `updateID` column for this table. For example, "survey11Add" for Survey 11.}
}
\description{
This function creates the TECH database table from raw survey data.
}
