% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeUpdateMetadata.R
\name{makeUpdateMetadata}
\alias{makeUpdateMetadata}
\title{Make the UPDATE_METADATA table}
\usage{
makeUpdateMetadata(updateID, date, updater, description)
}
\arguments{
\item{updateID}{Character string that will become the `updateID` column for this table. For example, "survey11Add" for Survey 11.}

\item{date}{Date of the update, in the format YYYY-MM-DD}

\item{updater}{Name of the person doing the update (first and last), e.g. Kaija Gahm}

\item{description}{Description of the update. For a standard database update, it's just "added [surveyID]".}
}
\description{
Make the UPDATE_METADATA table
}
