% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nsw}
\alias{nsw}
\title{Dialect regions: north/south/west}
\format{
An object of class \code{"sf", "data.frame"}.
}
\usage{
data(nsw)
}
\description{
Dialect region polygons for a rough division of the country into north, south, and west regions
}
\examples{
data(nsw)
}
\keyword{datasets}
