% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessingWrapper.R
\name{preprocessData}
\alias{preprocessData}
\title{Preprocess data}
\usage{
preprocessData(df, updateID)
}
\arguments{
\item{df}{Raw survey data as a data frame. Must have column 'responseID'.}

\item{updateID}{Character string that will become the `updateID` column for this table. For example, "survey11Add" for Survey 11.}
}
\description{
Wrapper function that combines `fixNames` and `worcesterRemove` into a single function
}
