% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessingFuns.R
\name{saveQuestionIDs}
\alias{saveQuestionIDs}
\title{Save questionID's}
\usage{
saveQuestionIDs(df)
}
\arguments{
\item{df}{Raw survey data as a data frame.}
}
\description{
This function separates out questionID's and column names from the raw survey data and returns them, for use in other functions.
}
