% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uaCols}
\alias{uaCols}
\title{US urban areas shapefiles (non-geometry columns only)}
\format{
A data frame/sf object with 3601 rows and 12 variables:
}
\source{
\url{https://hub.arcgis.com/datasets/1eb393576fb748fc9c4cc445018b30f2_0}
}
\usage{
uaCols
}
\description{
An sf object containing urban/rural area information to go with uaGeom. Info from the US census.
}
\keyword{datasets}
