% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uaGeom}
\alias{uaGeom}
\title{US urban areas shapefiles (geometry column only)}
\format{
A data frame/sf object with 3601 rows and 1 variable:
}
\source{
\url{https://hub.arcgis.com/datasets/1eb393576fb748fc9c4cc445018b30f2_0}
}
\usage{
uaGeom
}
\description{
An sf object containing geometry information to go with uaCols. Geometry only.
}
\keyword{datasets}
