% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appendTables.R
\name{updateTable}
\alias{updateTable}
\title{Append a newly-created table to the existing database table}
\usage{
updateTable(tableName, tableNew, overwrite = F, cn = con, u = updateIDString)
}
\arguments{
\item{tableName}{A character string: name of the database table. Not case sensitive. E.g. "responses" for the RESPONSES table.}

\item{tableNew}{The name of a data frame object (not in quotes), corresponding to the newly-created portion of the table in question. E.g. responsesNew.}

\item{overwrite}{FALSE by default. If FALSE, the function checks whether parameter `u` is already contained in the updateID column of the database table. If it's not, then the function appends the data from `tableNew`. If it is, the function does nothing and returns the table as is. If `overwrite` == TRUE, then if `u` is already in the database table, all data with updateID `u` will be *removed* and replaced with the data from `tableNew`.}

\item{cn}{SQLite connection to the database.}

\item{u}{the updateID you're using for this database update. When using the template script, `u` should be updateIDString.}
}
\description{
This function returns a complete version of the database table in question, as a data frame--i.e. the result contains both the rows from the new survey you are trying to add and all previous rows.
}
