% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worcesterRemove.R
\name{worcesterRemove}
\alias{worcesterRemove}
\title{Remove any obsolete geography columns}
\usage{
worcesterRemove(df)
}
\arguments{
\item{df}{Raw survey data as a data frame.}
}
\description{
Remove any obsolete geography columns
}
