% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anae}
\alias{anae}
\title{ANAE dialect regions}
\format{
A data frame/sf object with 17 rows and 6 variables:
\describe{
  \item{OBJECTID}{ID for GIS, I think?}
  \item{SHAPE_Leng}{Not sure what this is}
  \item{Dialect_Re}{Name of the dialect region}
  \item{Shape_Le_1}{Not sure what this is}
  \item{Shape_Area}{Area of the dialect region, not sure what units}
  \item{geometry}{sf geometry column}
  ...
}
}
\source{
YGDP-generated polygons
}
\usage{
anae
}
\description{
An sf object containing polygons delineating the dialect regions from the Atlas of North American English.
}
\keyword{datasets}
