% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carver}
\alias{carver}
\title{Carver dialect regions}
\format{
A data frame/sf object with 26 rows and 5 variables:
\describe{
  \item{Id}{ID for GIS, I think?}
  \item{Region}{Name of the region}
  \item{Sub_Region}{Name of the sub-region}
  \item{SubSub_Reg}{Name of the sub-sub-region}
  \item{geometry}{sf geometry column}
  ...
}
}
\source{
YGDP-generated polygons
}
\usage{
carver
}
\description{
An sf object containing polygons delineating the dialect regions defined in Carver (1987)
}
\keyword{datasets}
