% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counties}
\alias{counties}
\title{US counties shapefiles}
\format{
A data frame/sf object with 3142 rows and 56 variables:
}
\source{
\url{https://www.arcgis.com/home/item.html?id=a00d6b6149b34ed3b833e10fb72ef47b}
}
\usage{
counties
}
\description{
An sf object containing county polygons, from the US census
}
\keyword{datasets}
