% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counties}
\alias{counties}
\title{US census data by county}
\format{
An object of class \code{"sf", "data.frame"}.
}
\source{
https://www.arcgis.com/home/item.html?id=a00d6b6149b34ed3b833e10fb72ef47b
}
\usage{
data(counties)
}
\description{
County-level demographic data from the US census (2010)
}
\examples{
data(counties)

}
\keyword{datasets}
