% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessingFuns.R
\name{fixNames}
\alias{fixNames}
\title{Fix the names from the qualtrics table}
\usage{
fixNames(df)
}
\arguments{
\item{df}{Raw survey data as a data frame. Must have column 'responseID'.}
}
\description{
Since the Qualtrics output has a weird three-line header, this function cleans up the names. Sets the first row of the data frame as names, then removes the first and second rows of the data frame. Then it renames a bunch of still-messy names from Qualtrics.
}
