% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCensusCountyDemo.R
\name{makeCensusCountyDemo}
\alias{makeCensusCountyDemo}
\title{Make the CENSUS_COUNTY_DEMO table}
\usage{
makeCensusCountyDemo(cities, con, updateID)
}
\arguments{
\item{cities}{The CITIES element of the list returned by makeDemoGeo.R. It will be passed into this function to be further modified before being finalized.}

\item{con}{A connection to the existing database, created with RSQLite.}

\item{updateID}{Character string that will become the `updateID` column for this table. For example, "survey11Add" for Survey 11.}
}
\description{
Make the CENSUS_COUNTY_DEMO table
}
