% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSurveySentences.R
\name{makeSurveySentences}
\alias{makeSurveySentences}
\title{Make the SURVEY_SENTENCES table}
\usage{
makeSurveySentences(df, surveyID, updateID)
}
\arguments{
\item{df}{Raw survey data as a data frame.}

\item{surveyID}{Character string that will become the `surveyID` column for this table. For example, "S11" for Survey 11.}

\item{updateID}{Character string that will become the `updateID` column for this table. For example, "survey11Add" for Survey 11.}
}
\description{
Make the SURVEY_SENTENCES table
}
