% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTables.R
\name{makeVersionHistory}
\alias{makeVersionHistory}
\title{Make the VERSION_HISTORY table}
\usage{
makeVersionHistory(con, date, description)
}
\arguments{
\item{con}{A connection to the existing database, created with RSQLite.}

\item{date}{Date of the update, in the format YYYY-MM-DD}

\item{description}{Description of the version. For a standard database update, it's just "added [surveyID]".}
}
\description{
This function creates the row of the VERSION_HISTORY table that corresponds to the update we're making for this survey
}
