% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nsw}
\alias{nsw}
\title{NSW dialect regions}
\format{
A data frame/sf object with 3 rows and 5 variables:
\describe{
  \item{OBJECTID}{ID for GIS, I think?}
  \item{SHAPE_Leng}{I don't know what this is}
  \item{SHAPE_Area}{Area of the polygon}
  \item{Dialect_Re}{Name of the region (north, south, or west)}
  \item{geometry}{sf geometry column}
  ...
}
}
\source{
YGDP-generated polygons
}
\usage{
nsw
}
\description{
An sf object containing polygons dividing the country into North/South/West regions.
}
\keyword{datasets}
