% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ua}
\alias{ua}
\title{US census urban areas data}
\format{
An object of class \code{"sf", "data.frame"}.
}
\source{
https://www.census.gov/programs-surveys/geography/guidance/geo-areas/urban-rural/2010-urban-rural.html
}
\usage{
data(ua)
}
\description{
Urban area data from the US census
}
\examples{
data(ua)

}
\keyword{datasets}
