% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worcesterRemove.R
\name{worcesterRemove}
\alias{worcesterRemove}
\title{Remove any obsolete geography columns}
\usage{
worcesterRemove(df)
}
\arguments{
\item{df}{Raw survey data as a data frame.}
}
\description{
Before Kaija, Jason Zentz had rigged something up in Qualtrics that would geocode the localities--except it doesn't really work, and you end up with a lot of Worcester. This function removes the old geo stuff.
}
